（1）修正flv，f4v 无法得到帧数的bug。 （2）目前这个版本只支持帧率为整数，对帧率为有理数的视频所得到的帧率和帧数有偏差。

+++ src/Geb.Video.FFMPEG/VideoFileReader.h
@@ -219,9 +219,9 @@ namespace Geb { namespace Video { namespace FFMPEG
 
 		array<Byte>^ ReadAudioFrame( bool onlyCurrentVideoFrame );
 
-		int Seek(long long frameIndex,  Boolean seekKeyFrame);
+		long long Seek(long long frameIndex,  Boolean seekKeyFrame);
 
-		int Seek(long long frameIndex)
+		long long Seek(long long frameIndex)
 		{
 			return Seek(frameIndex, true);
 		}
